digital_twin/requirements.txt
# Core ML/AI
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
axolotl==0.4.1
torch==2.5.0
transformers==4.46.0
peft==0.12.0
trl==0.9.6
datasets==3.0.2
accelerate==1.1.1
bitsandbytes==0.44.0

# RAG & Embeddings
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface==0.1.0
chromadb==0.5.0
sentence-transformers==3.1.1

# API & Server
fastapi==0.115.0
uvicorn==0.32.0
slowapi==0.1.9
pydantic==2.9.0
python-multipart==0.0.9

# Inference
vllm==0.6.3
ollama==0.3.0

# Gmail Integration
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Data Processing
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.1

# Security & Privacy
presidio-analyzer==2.1.5
presidio-anonymizer==2.1.5

# Evaluation
nltk==3.9.0
scikit-learn==1.5.1

# UI & Testing
streamlit==1.39.0
pytest==8.3.3
pytest-asyncio==0.23.6
pytest-cov==5.0.0

# Utilities
python-dotenv==1.0.1
tqdm==4.66.4